extract common activate/update request assembly methods into a component
