+++ internal/cli/command/cluster/feature/features/securityscan/update.go
@@ -18,11 +18,9 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"strings"
 
 	"emperror.dev/errors"
 	"github.com/AlecAivazis/survey/v2"
-	"github.com/antihax/optional"
 	"github.com/banzaicloud/banzai-cli/.gen/pipeline"
 	"github.com/banzaicloud/banzai-cli/internal/cli"
 	clustercontext "github.com/banzaicloud/banzai-cli/internal/cli/command/cluster/context"
@@ -34,6 +32,7 @@ import (
 
 func NewUpdateCommand(banzaiCli cli.Cli) *cobra.Command {
 	options := updateOptions{}
+	fu := MakeFeatureUpdater(banzaiCli)
 
 	cmd := &cobra.Command{
 		Use:     "update",
@@ -41,7 +40,7 @@ func NewUpdateCommand(banzaiCli cli.Cli) *cobra.Command {
 		Short:   fmt.Sprintf("Update the %s feature of a cluster", featureName),
 		Args:    cobra.NoArgs,
 		RunE: func(_ *cobra.Command, args []string) error {
-			return runUpdate(banzaiCli, options, args)
+			return fu.runUpdate(options, args)
 		},
 	}
 
@@ -60,7 +59,7 @@ type updateOptions struct {
 
 // featureUpdater struct for gathering helper operations for the feature update
 type featureUpdater struct {
-	banzaiCLI cli.Cli
+	specAssembler
 }
 
 func MakeFeatureUpdater(banzaiCLI cli.Cli) *featureUpdater {
@@ -76,11 +75,6 @@ func (fu *featureUpdater) getSecurityScanFeature(orgID int32, clusterID int32) (
 		return nil, errors.WrapIf(err, "failed to retrieve the feature to update")
 	}
 
-	//var securityFeatureSpec *SecurityScanFeatureSpec
-	//if err := mapstructure.Decode(clusterFeatureDetails.Spec, securityFeatureSpec); err != nil {
-	//	return nil, errors.WrapIf(err, "failed to decode the feature to update")
-	//}
-
 	return clusterFeatureDetails.Spec, nil
 }
 
@@ -93,14 +87,12 @@ func ValidateUpdateClusterFeatureRequest(req interface{}) error {
 	return nil
 }
 
-func runUpdate(banzaiCli cli.Cli, options updateOptions, args []string) error {
+func (fu *featureUpdater) runUpdate(options updateOptions, args []string) error {
 	if err := options.Init(args...); err != nil {
 		return errors.Wrap(err, "failed to initialize options")
 	}
 
-	fu := MakeFeatureUpdater(banzaiCli)
-
-	orgId := banzaiCli.Context().OrganizationID()
+	orgId := fu.banzaiCLI.Context().OrganizationID()
 	clusterId := options.ClusterID()
 
 	securityScanFeatureSpec, err := fu.getSecurityScanFeature(orgId, clusterId)
@@ -111,19 +103,19 @@ func runUpdate(banzaiCli cli.Cli, options updateOptions, args []string) error {
 	req := new(pipeline.UpdateClusterFeatureRequest)
 	req.Spec = securityScanFeatureSpec
 
-	if options.filePath == "" && banzaiCli.Interactive() {
+	if options.filePath == "" && fu.banzaiCLI.Interactive() {
 		if err := fu.buildUpdateReqInteractively(options, req); err != nil {
 			return errors.WrapIf(err, "failed to build update request interactively")
 		}
-	} else {
-		if err := readUpdateReqFromFileOrStdin(options.filePath, req); err != nil {
-			return errors.WrapIff(err, "failed to read %s cluster feature specification", featureName)
-		}
 	}
 
-	resp, err := banzaiCli.Client().ClusterFeaturesApi.UpdateClusterFeature(context.Background(), orgId, clusterId, featureName, *req)
+	if err := readUpdateReqFromFileOrStdin(options.filePath, req); err != nil {
+		return errors.WrapIff(err, "failed to read %s cluster feature specification", featureName)
+	}
+
+	resp, err := fu.banzaiCLI.Client().ClusterFeaturesApi.UpdateClusterFeature(context.Background(), orgId, clusterId, featureName, *req)
 	if err != nil {
-		cli.LogAPIError(fmt.Sprintf("activate %s cluster feature", featureName), err, resp.Request)
+		cli.LogAPIError(fmt.Sprintf("activate %s cluster feature", featureName), err, resp)
 		log.Fatalf("could not activate %s cluster feature: %v", featureName, err)
 	}
 
@@ -147,9 +139,13 @@ func readUpdateReqFromFileOrStdin(filePath string, req *pipeline.UpdateClusterFe
 
 func (fu *featureUpdater) buildUpdateReqInteractively(_ updateOptions, req *pipeline.UpdateClusterFeatureRequest) error {
 	var edit bool
-	if err := survey.AskOne(&survey.Confirm{Message: "Edit the cluster feature update request in your text editor?"}, &edit); err != nil {
+	if err := survey.AskOne(
+		&survey.Confirm{Message: "Edit the cluster feature update request in your text editor?"},
+		&edit,
+		survey.WithValidator(ValidateUpdateClusterFeatureRequest)); err != nil {
 		return errors.WrapIf(err, "failure during survey")
 	}
+
 	if !edit {
 		return fu.buildCustomAnchoreFeatureRequest(req)
 	}
@@ -165,6 +161,7 @@ func (fu *featureUpdater) buildUpdateReqInteractively(_ updateOptions, req *pipe
 	if err := survey.AskOne(prompt, &updated); err != nil {
 		return errors.WrapIf(err, "failure during survey")
 	}
+
 	if err := json.Unmarshal([]byte(updated), req); err != nil {
 		return errors.WrapIf(err, "failed to unmarshal JSON as request")
 	}
@@ -180,18 +177,18 @@ func (fu *featureUpdater) buildCustomAnchoreFeatureRequest(updateRequest *pipeli
 		return errors.WrapIf(err, "failed to decode the feature to update")
 	}
 
-	anchoreConfig, err := fu.askForAnchoreConfig(securityFeatureSpec.CustomAnchore)
+	anchoreConfig, err := fu.askForAnchoreConfig(&securityFeatureSpec.CustomAnchore)
 	if err != nil {
 		return errors.WrapIf(err, "failed to read Anchore configuration details")
 	}
 
-	policy, err := fu.askForPolicy(securityFeatureSpec.Policy)
+	policy, err := fu.askForPolicy(&securityFeatureSpec.Policy)
 	if err != nil {
 		return errors.WrapIf(err, "failed to read Anchore Policy configuration details")
 	}
 
 	// todo whitelist updates not supported fro now
-	webhookConfig, err := fu.askForWebHookConfig(securityFeatureSpec.WebhookConfig)
+	webhookConfig, err := fu.askForWebHookConfig(&securityFeatureSpec.WebhookConfig)
 	if err != nil {
 		return errors.WrapIf(err, "failed to read webhook configuration")
 	}
@@ -207,200 +204,4 @@ func (fu *featureUpdater) buildCustomAnchoreFeatureRequest(updateRequest *pipeli
 	updateRequest.Spec = ssfMap
 
 	return nil
-}
-
-func (fu *featureUpdater) askForAnchoreConfig(currentSpec anchoreSpec) (*anchoreSpec, error) {
-
-	var customAnchore bool
-	if err := survey.AskOne(
-		&survey.Confirm{
-			Message: "Configure a custom anchore instance? ",
-		},
-		&customAnchore,
-	); err != nil {
-		return nil, errors.WrapIf(err, "failure during survey")
-	}
-
-	if !customAnchore {
-		return &anchoreSpec{
-			Enabled: false,
-		}, nil
-	}
-
-	// custom anchore config
-	var anchoreURL string
-	if err := survey.AskOne(
-		&survey.Input{
-			Message: "Please enter the custom anchore URL:",
-			Default: currentSpec.Url,
-		},
-		&anchoreURL,
-	); err != nil {
-		return nil, errors.WrapIf(err, "failed to read custom Anchore URL")
-	}
-
-	secretID, err := fu.askForSecret(currentSpec.SecretID)
-	if err != nil {
-		return nil, errors.WrapIf(err, "failed to read secret for accessing custom Anchore ")
-	}
-
-	return &anchoreSpec{
-		Enabled:  true,
-		Url:      anchoreURL,
-		SecretID: secretID,
-	}, nil
-}
-
-func (fu *featureUpdater) askForSecret(secretID string) (string, error) {
-	const (
-		PasswordSecretType = "password"
-	)
-
-	orgID := fu.banzaiCLI.Context().OrganizationID()
-
-	secrets, _, err := fu.banzaiCLI.Client().SecretsApi.GetSecrets(context.Background(), orgID, &pipeline.GetSecretsOpts{Type_: optional.NewString(PasswordSecretType)})
-	if err != nil {
-		return "", errors.Wrap(err, "failed to retrieve secrets")
-	}
-
-	// TODO add create secret option
-	if len(secrets) == 0 {
-		return "", errors.New(fmt.Sprintf("there are no secrets with type %q", PasswordSecretType))
-	}
-
-	options := make([]string, len(secrets))
-	currentSecretName := ""
-	for i, s := range secrets {
-		options[i] = s.Name
-		if s.Id == secretID {
-			currentSecretName = s.Name
-		}
-	}
-
-	var secretName string
-	if err := survey.AskOne(
-		&survey.Select{
-			Message: "Please select a secret to access the custom Anchore instance:",
-			Options: options,
-			Default: currentSecretName,
-		},
-		&secretName,
-	); err != nil {
-		return "", errors.WrapIf(err, "failed to retrieve secrets")
-	}
-
-	for _, s := range secrets {
-		if s.Name == secretName {
-			return s.Id, nil
-		}
-	}
-
-	return "", errors.Errorf("no secret with name %q", secretName)
-}
-
-func (fu *featureUpdater) askForPolicy(currentPolicy policySpec) (*policySpec, error) {
-	type policy struct {
-		name string
-		id   string
-	}
-	// todo add all supported policies here
-	policies := []policy{{"Default bundle", "1"}}
-
-	options := make([]string, len(policies))
-	currentPolicyName := ""
-	for i, s := range policies {
-		options[i] = s.name
-		if s.id == currentPolicy.PolicyID {
-			currentPolicyName = s.name
-		}
-	}
-
-	var policyName string
-	if err := survey.AskOne(
-		&survey.Select{
-			Message: "Please select a policy for the Anchor Engine:",
-			Options: options,
-			Default: currentPolicyName,
-		},
-		&policyName,
-	); err != nil {
-		return nil, errors.WrapIf(err, "failed to select policy")
-	}
-
-	var selected policy
-	for _, s := range policies {
-		if s.name == policyName {
-			selected = s
-		}
-	}
-
-	return &policySpec{
-		PolicyID: selected.id,
-	}, nil
-}
-
-func (fu *featureUpdater) askForWebHookConfig(currentWH webHookConfigSpec) (*webHookConfigSpec, error) {
-	var enable bool
-	if err := survey.AskOne(
-		&survey.Confirm{
-			Message: "Configure the security scan webhook?",
-		},
-		&enable,
-	); err != nil {
-		return nil, errors.WrapIf(err, "failure during survey")
-	}
-
-	if !enable {
-		return &webHookConfigSpec{
-			Enabled: false,
-		}, nil
-	}
-
-	var selector string
-	if err := survey.AskOne(
-		&survey.Select{
-			Message: "Please choose the selector for the webhook configuration:",
-			Options: []string{"Exclude", "Include"},
-			Default: currentWH.Selector,
-		},
-		&selector,
-	); err != nil {
-		return nil, errors.WrapIf(err, "failed to select policy")
-	}
-
-	var namespaces string
-	if err := survey.AskOne(
-		&survey.Input{
-			Message: "Please enter the comma separated list of namespaces:",
-			Default: strings.Join(currentWH.Namespaces, ","),
-		},
-		&namespaces,
-	); err != nil {
-		return nil, errors.WrapIf(err, "failed to read namespaces")
-	}
-
-	return &webHookConfigSpec{
-		Enabled:    true,
-		Selector:   selector,
-		Namespaces: strings.Split(namespaces, ","),
-	}, nil
-}
-
-func (fu *featureUpdater) securityScanSpecAsMap(spec *SecurityScanFeatureSpec) (map[string]interface{}, error) {
-	// fill the structure of the config - make filling up the values easier
-	if spec == nil {
-		spec = &SecurityScanFeatureSpec{
-			CustomAnchore:    anchoreSpec{},
-			Policy:           policySpec{},
-			ReleaseWhiteList: nil,
-			WebhookConfig:    webHookConfigSpec{},
-		}
-	}
-
-	var specMap map[string]interface{}
-	if err := mapstructure.Decode(spec, &specMap); err != nil {
-		return nil, err
-	}
-
-	return specMap, nil
 }
